Fixed some issues with beans and controllers

+++ src/it/beije/erp/timesheet/controller/AddressController.java
@@ -1,14 +1,9 @@
 package it.beije.erp.timesheet.controller;
 
 import java.sql.Date;
-import java.time.LocalDate;
-import java.util.ArrayList;
+
 import java.util.Calendar;
 import java.util.List;
-
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.Persistence;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -16,7 +11,6 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
 
 import it.beije.erp.entity.Address;
 import it.beije.erp.entity.User;
@@ -27,7 +21,7 @@ public class AddressController {
 	
 	// GET : read addresses from User within id
 	@RequestMapping(value = "/addresses/{id}", method = RequestMethod.GET)
-	public String getAddresses(@PathVariable int id, Model model) {
+	public String getAddresses(@PathVariable Long id, Model model) {
 		
 		User user = JPAService.getBean(User.class, id);
 		List<Address> addresses = user.getAddresses();
